update wsscxf fat ee7ee8split with conditional check
